requirements.txt
distlib==0.3.9
filelock==3.18.0
platformdirs==4.3.6
virtualenv==20.29.3
django
gunicorn
psycopg2
python-dotenv
